asdfasf

+++ TrashGame/src/simplePhysics/Area.java
@@ -9,7 +9,7 @@ public class Area extends Rectangle2D.Float{
     protected boolean isActive = false;
     protected boolean isMovable = false;
     public ID id = ID.AREA;
-    public Area(float x, float y, int width, int height){
+    public Area(float x, float y, float width, float height){
         super(x, y, width, height);
     }
